project-nanda/requirements.txt
nest @ git+https://github.com/projnanda/NEST.git@v1
requests
python-dotenv
pyngrok
